+++ client/src/components/LibraryNavigator/LibraryDataProvider.ts
@@ -1,22 +1,48 @@
 // Copyright © 2023, SAS Institute Inc., Cary, NC, USA.  All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import { Writable } from "stream";
 import {
+  CancellationToken,
+  DataTransfer,
+  DataTransferItem,
   Disposable,
+  DocumentDropEdit,
+  DocumentSelector,
   Event,
   EventEmitter,
+  Position,
   ProviderResult,
+  TextDocument,
   TreeDataProvider,
+  TreeDragAndDropController,
   TreeItem,
   TreeItemCollapsibleState,
+  TreeView,
   Uri,
+  languages,
+  window,
 } from "vscode";
+import { SubscriptionProvider } from "../SubscriptionProvider";
 import LibraryModel from "./LibraryModel";
 import { Icons, Messages, WorkLibraryId } from "./const";
 import { LibraryItem, LibraryType, TableType } from "./types";
 
-class LibraryDataProvider implements TreeDataProvider<LibraryItem> {
+export const libraryItemMimeType =
+  "application/vnd.code.tree.librarydataprovider";
+const tableTextMimeType = `${libraryItemMimeType}.text`;
+class LibraryDataProvider
+  implements
+    TreeDataProvider<LibraryItem>,
+    TreeDragAndDropController<LibraryItem>,
+    SubscriptionProvider
+{
   private _onDidChangeTreeData = new EventEmitter<LibraryItem | undefined>();
+  private _treeView: TreeView<LibraryItem>;
+  private _dropEditProvider: Disposable;
+
+  public dropMimeTypes: string[] = [];
+  public dragMimeTypes: string[] = [libraryItemMimeType, tableTextMimeType];
 
   get onDidChangeTreeData(): Event<LibraryItem> {
     return this._onDidChangeTreeData.event;
@@ -25,7 +51,53 @@ class LibraryDataProvider implements TreeDataProvider<LibraryItem> {
   constructor(
     private readonly model: LibraryModel,
     private readonly extensionUri: Uri,
-  ) {}
+  ) {
+    this._treeView = window.createTreeView("librarydataprovider", {
+      treeDataProvider: this,
+      dragAndDropController: this,
+      canSelectMany: true,
+    });
+    this._dropEditProvider = languages.registerDocumentDropEditProvider(
+      this.selector(),
+      this,
+    );
+  }
+
+  public getSubscriptions(): Disposable[] {
+    return [this._treeView, this._dropEditProvider];
+  }
+
+  public handleDrag(
+    source: LibraryItem[],
+    dataTransfer: DataTransfer,
+  ): void | Thenable<void> {
+    const dataTransferItem = new DataTransferItem(source);
+    dataTransfer.set(libraryItemMimeType, dataTransferItem);
+    if (source?.[0].library) {
+      dataTransfer.set(
+        tableTextMimeType,
+        new DataTransferItem(source?.[0].uid),
+      );
+    }
+  }
+
+  public async provideDocumentDropEdits(
+    _document: TextDocument,
+    position: Position,
+    dataTransfer: DataTransfer,
+    token: CancellationToken,
+  ): Promise<DocumentDropEdit | undefined> {
+    const dataTransferItem = dataTransfer.get(this.dragMimeTypes[1]);
+    if (token.isCancellationRequested || !dataTransferItem) {
+      return undefined;
+    }
+
+    return { insertText: dataTransferItem.value };
+  }
+
+  public selector(): DocumentSelector {
+    return { language: "sas" };
+  }
 
   public getTreeItem(item: LibraryItem): TreeItem | Promise<TreeItem> {
     const iconPath = this.iconPathForItem(item);
@@ -47,7 +119,11 @@ class LibraryDataProvider implements TreeDataProvider<LibraryItem> {
         item.type === TableType
           ? {
               command: "SAS.viewTable",
-              arguments: [item, this.model.getTableResultSet(item)],
+              arguments: [
+                item,
+                this.model.getTableResultSet(item),
+                () => this.model.fetchColumns(item),
+              ],
               title: Messages.ViewTableCommandTitle,
             }
           : undefined,
@@ -77,6 +153,10 @@ class LibraryDataProvider implements TreeDataProvider<LibraryItem> {
 
   public getChildren(item?: LibraryItem): ProviderResult<LibraryItem[]> {
     return this.model.getChildren(item);
+  }
+
+  public writeTableContentsToStream(stream: Writable, item: LibraryItem) {
+    return this.model.writeTableContentsToStream(stream, item);
   }
 
   public async deleteTable(item: LibraryItem): Promise<void> {