+++ client/src/components/LibraryNavigator/index.ts
@@ -1,19 +1,19 @@
 // Copyright © 2023, SAS Institute Inc., Cary, NC, USA.  All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+import { createWriteStream } from "fs";
 import {
   commands,
   ConfigurationChangeEvent,
   Disposable,
   ExtensionContext,
-  TreeView,
   Uri,
   window,
   workspace,
 } from "vscode";
+import { Column } from "../../connection/rest/api/compute";
 import DataViewer from "../../panels/DataViewer";
 import { WebViewManager } from "../../panels/WebviewManager";
-import DragAndDropController from "../DragAndDropController";
 import { SubscriptionProvider } from "../SubscriptionProvider";
 import LibraryDataProvider from "./LibraryDataProvider";
 import LibraryModel from "./LibraryModel";
@@ -22,36 +22,35 @@ import { LibraryItem, TableData } from "./types";
 
 class LibraryNavigator implements SubscriptionProvider {
   private libraryDataProvider: LibraryDataProvider;
-  private treeView: TreeView<LibraryItem>;
   private extensionUri: Uri;
   private webviewManager: WebViewManager;
 
   constructor(context: ExtensionContext) {
     this.extensionUri = context.extensionUri;
-    const dragAndDropController = new DragAndDropController<LibraryItem>(
-      context,
-      "application/vnd.code.tree.sas-library-navigator",
-      (item: LibraryItem | undefined) => (item.library ? item.uid : undefined),
-    );
     this.libraryDataProvider = new LibraryDataProvider(
       new LibraryModel(),
       context.extensionUri,
     );
-    this.treeView = window.createTreeView("sas-library-navigator", {
-      treeDataProvider: this.libraryDataProvider,
-      dragAndDropController,
-    });
     this.webviewManager = new WebViewManager();
   }
 
   public getSubscriptions(): Disposable[] {
     return [
-      this.treeView,
+      ...this.libraryDataProvider.getSubscriptions(),
       commands.registerCommand(
         "SAS.viewTable",
-        async (item: LibraryItem, paginator: PaginatedResultSet<TableData>) => {
+        async (
+          item: LibraryItem,
+          paginator: PaginatedResultSet<TableData>,
+          fetchColumns: () => Column[],
+        ) => {
           this.webviewManager.render(
-            new DataViewer(this.extensionUri, item.uid, paginator),
+            new DataViewer(
+              this.extensionUri,
+              item.uid,
+              paginator,
+              fetchColumns,
+            ),
             item.uid,
           );
         },
@@ -64,9 +63,29 @@ class LibraryNavigator implements SubscriptionProvider {
           window.showErrorMessage(error.message);
         }
       }),
+      commands.registerCommand(
+        "SAS.downloadTable",
+        async (item: LibraryItem) => {
+          const uri = await window.showSaveDialog({
+            defaultUri: Uri.file(
+              `${item.library}.${item.name}.csv`.toLocaleLowerCase(),
+            ),
+          });
+
+          if (!uri) {
+            return;
+          }
+
+          const stream = createWriteStream(uri.fsPath);
+          await this.libraryDataProvider.writeTableContentsToStream(
+            stream,
+            item,
+          );
+        },
+      ),
       commands.registerCommand("SAS.collapseAllLibraries", () => {
         commands.executeCommand(
-          "workbench.actions.treeView.sas-library-navigator.collapseAll",
+          "workbench.actions.treeView.librarydataprovider.collapseAll",
         );
       }),
       workspace.onDidChangeConfiguration((event: ConfigurationChangeEvent) => {