+++ client/src/webview/DataViewer.tsx
@@ -3,57 +3,50 @@
 
 import ".";
 
-import {
-  VSCodeDataGrid,
-  VSCodeDataGridCell,
-  VSCodeDataGridRow,
-} from "@vscode/webview-ui-toolkit/react";
-import * as React from "react";
-import * as ReactDOMClient from "react-dom/client";
-import InfiniteScroll from "react-infinite-scroller";
+import { AgGridReact } from "ag-grid-react";
+import { createElement, useState } from "react";
+import { createRoot } from "react-dom/client";
 import useDataViewer from "./useDataViewer";
 
+import "ag-grid-community/styles/ag-grid.css";
+import "ag-grid-community/styles/ag-theme-alpine.css";
+import "./DataViewer.css";
+
+const gridStyles = {
+  "--ag-borders": "none",
+  "--ag-row-border-width": "0px",
+  height: "100%",
+  width: "100%",
+};
+
 const DataViewer = () => {
-  const { loadMoreResults, headers, rows, hasMore } = useDataViewer();
+  const { columns, onGridReady } = useDataViewer();
+  const [theme] = useState(
+    document.querySelector(".vscode-dark")
+      ? "ag-theme-alpine-dark"
+      : "ag-theme-alpine",
+  );
+
+  if (columns.length === 0) {
+    return null;
+  }
 
   return (
-    <div>
-      <InfiniteScroll
-        pageStart={0}
-        loadMore={loadMoreResults}
-        hasMore={hasMore}
-      >
-        <VSCodeDataGrid>
-          <VSCodeDataGridRow row-type="header">
-            {(headers.columns || []).map((column, idx) => (
-              <VSCodeDataGridCell
-                style={{ minWidth: "100px" }}
-                cell-type="columnheader"
-                key={idx}
-                grid-column={idx + 1}
-              >
-                {column}
-              </VSCodeDataGridCell>
-            ))}
-          </VSCodeDataGridRow>
-          {rows.map((row, idx) => (
-            <VSCodeDataGridRow key={idx}>
-              {row.cells.map((cell, idx) => (
-                <VSCodeDataGridCell
-                  key={idx}
-                  grid-column={idx + 1}
-                  style={{ minWidth: "100px" }}
-                >
-                  {cell}
-                </VSCodeDataGridCell>
-              ))}
-            </VSCodeDataGridRow>
-          ))}
-        </VSCodeDataGrid>
-      </InfiniteScroll>
+    <div className={`ag-grid-wrapper ${theme}`} style={gridStyles}>
+      <AgGridReact
+        cacheBlockSize={100}
+        columnDefs={columns}
+        defaultColDef={{
+          resizable: true,
+        }}
+        infiniteInitialRowCount={100}
+        maxBlocksInCache={10}
+        onGridReady={onGridReady}
+        rowModelType="infinite"
+      />
     </div>
   );
 };
 
-const root = ReactDOMClient.createRoot(document.querySelector(".data-viewer"));
+const root = createRoot(document.querySelector(".data-viewer"));
 root.render(<DataViewer />);