+++ client/src/webview/useDataViewer.ts
@@ -1,7 +1,12 @@
 // Copyright © 2023, SAS Institute Inc., Cary, NC, USA.  All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import { useEffect, useState } from "react";
+import { ColDef, GridReadyEvent, IGetRowsParams } from "ag-grid-community";
+import { useCallback, useEffect, useState } from "react";
+import { v4 } from "uuid";
+import { TableData } from "../components/LibraryNavigator/types";
+import { Column } from "../connection/rest/api/compute";
+import columnHeaderTemplate from "./columnHeaderTemplate";
 
 declare const acquireVsCodeApi;
 const vscode = acquireVsCodeApi();
@@ -10,69 +15,140 @@ const contextMenuHandler = (e) => {
   e.stopImmediatePropagation();
 };
 
-const useDataViewer = () => {
-  const [headers, setHeaders] = useState({});
-  const [rows, setRows] = useState([]);
-  const [hasMore, setHasMore] = useState(false);
-  const [start, setStart] = useState(0);
+const defaultTimeout = 60 * 1000; // 60 seconds (accounting for compute session expiration)
 
-  useEffect(() => {
-    if (rows.length === 0) {
-      return;
-    }
-    vscode.setState({ headers, rows, hasMore, start });
-  }, [headers, rows, hasMore, start]);
+let queryTableDataTimeoutId: ReturnType<typeof setTimeout> | null = null;
+const clearQueryTimeout = (): void => {
+  if (!queryTableDataTimeoutId) {
+    return;
+  }
+  clearTimeout(queryTableDataTimeoutId);
+  queryTableDataTimeoutId = null;
+};
+const queryTableData = (start: number, end: number): Promise<TableData> => {
+  const requestKey = v4();
+  vscode.postMessage({
+    command: "request:loadData",
+    key: requestKey,
+    data: { start, end },
+  });
 
-  const commandHandler = (event) => {
-    const { data } = event.data;
-    switch (event.data.command) {
-      case "response:loadData":
-        setHeaders(data.headers);
-        setRows(data.rows);
-        setHasMore(data.hasMore);
-        setStart(data.start);
-        break;
-      case "response:loadMoreResults": {
-        setRows((rows) => rows.concat(event.data.data.rows));
-        setHasMore(data.hasMore);
-        setStart(data.start);
-        break;
+  return new Promise((resolve, reject) => {
+    const commandHandler = (event) => {
+      const { data } = event.data;
+      if (event.data.key !== requestKey) {
+        return;
       }
-      default:
-        break;
-    }
-  };
+      if (event.data.command === "response:loadData") {
+        window.removeEventListener("message", commandHandler);
+        clearQueryTimeout();
+        resolve(data);
+      }
+    };
 
-  const loadMoreResults = () =>
-    vscode.postMessage({ command: "request:loadMoreResults" });
-
-  useEffect(() => {
-    const serializedState = vscode.getState();
-    // If we have serialized data, lets initialize our component with that
-    // data and update the start offset so the paginator knows the next data
-    // to pull in.
-    if (serializedState) {
-      const { headers, rows, hasMore, start } = serializedState;
-      setHeaders(headers);
-      setRows(rows);
-      setHasMore(hasMore);
-      setStart(start);
-      vscode.postMessage({ command: "request:updateStart", data: { start } });
-    } else {
-      vscode.postMessage({ command: "request:loadData" });
-    }
+    clearQueryTimeout();
+    queryTableDataTimeoutId = setTimeout(() => {
+      window.removeEventListener("message", commandHandler);
+      reject(new Error("Timeout exceeded"));
+    }, defaultTimeout);
 
     window.addEventListener("message", commandHandler);
+  });
+};
 
+let fetchColumnsTimeoutId: ReturnType<typeof setTimeout> | null = null;
+const clearFetchColumnsTimeout = () =>
+  fetchColumnsTimeoutId && clearTimeout(fetchColumnsTimeoutId);
+const fetchColumns = (): Promise<Column[]> => {
+  const requestKey = v4();
+  vscode.postMessage({ command: "request:loadColumns", key: requestKey });
+
+  return new Promise((resolve, reject) => {
+    const commandHandler = (event) => {
+      const { data } = event.data;
+      if (event.data.key !== requestKey) {
+        return;
+      }
+      if (event.data.command === "response:loadColumns") {
+        window.removeEventListener("message", commandHandler);
+        clearFetchColumnsTimeout();
+        resolve(data);
+      }
+    };
+
+    clearFetchColumnsTimeout();
+    fetchColumnsTimeoutId = setTimeout(() => {
+      window.removeEventListener("message", commandHandler);
+      reject(new Error("Timeout exceeded"));
+    }, defaultTimeout);
+
+    window.addEventListener("message", commandHandler);
+  });
+};
+
+const useDataViewer = () => {
+  const [columns, setColumns] = useState<ColDef[]>([]);
+
+  const onGridReady = useCallback(
+    (event: GridReadyEvent) => {
+      const dataSource = {
+        rowCount: undefined,
+        getRows: (params: IGetRowsParams) => {
+          queryTableData(params.startRow, params.endRow).then(
+            ({ rows, count }: TableData) => {
+              const rowData = rows.map(({ cells }) => {
+                const row = cells.reduce(
+                  (carry, cell, index) => ({
+                    ...carry,
+                    [columns[index].field]: cell,
+                  }),
+                  {},
+                );
+
+                return row;
+              });
+
+              params.successCallback(rowData, count);
+            },
+          );
+        },
+      };
+
+      event.api.setDatasource(dataSource);
+    },
+    [columns],
+  );
+
+  useEffect(() => {
+    if (columns.length > 0) {
+      return;
+    }
+
+    fetchColumns().then((columnsData) => {
+      const columns: ColDef[] = columnsData.map((column) => ({
+        field: column.name,
+        headerComponentParams: {
+          template: columnHeaderTemplate(column.type),
+        },
+      }));
+      columns.unshift({
+        field: "#",
+        suppressMovable: true,
+      });
+
+      setColumns(columns);
+    });
+  }, [columns.length]);
+
+  useEffect(() => {
     window.addEventListener("contextmenu", contextMenuHandler, true);
 
     return () => {
-      window.removeEventListener("message", commandHandler);
       window.removeEventListener("contextmenu", contextMenuHandler);
     };
   }, []);
 
-  return { loadMoreResults, headers, rows, hasMore };
+  return { columns, onGridReady };
 };
 
 export default useDataViewer;