+++ client/test/index.ts
@@ -1,6 +1,6 @@
-import * as path from "path";
-import * as Mocha from "mocha";
-import * as glob from "glob";
+import glob from "glob";
+import Mocha from "mocha";
+import path from "path";
 
 export function run(): Promise<void> {
   // Create the mocha test