+++ tools/build.mjs
@@ -1,6 +1,8 @@
+import concurrently from "concurrently";
+import esbuild from "esbuild";
+
 console.log("start");
 const dev = process.argv[2];
-import esbuild from "esbuild";
 
 const plugins = [
   {
@@ -15,19 +17,14 @@ const plugins = [
   },
 ];
 
-const buildOptions = {
-  entryPoints: {
-    "./client/dist/node/extension": "./client/src/node/extension.ts",
-    "./server/dist/node/server": "./server/src/node/server.ts",
-    "./client/dist/webview/DataViewer": "./client/src/webview/DataViewer.tsx",
-  },
+const commonBuildOptions = {
   bundle: true,
   outdir: ".",
-  platform: "node",
   external: ["vscode"],
   loader: {
     ".properties": "text",
     ".node": "copy",
+    ".svg": "dataurl",
   },
   sourcemap: !!dev,
   minify: !dev,
@@ -37,11 +34,47 @@ const buildOptions = {
   },
 };
 
-const ctx = await esbuild.context(buildOptions);
-await ctx.rebuild();
+const nodeBuildOptions = {
+  ...commonBuildOptions,
+  entryPoints: {
+    "./client/dist/node/extension": "./client/src/node/extension.ts",
+    "./server/dist/node/server": "./server/src/node/server.ts",
+  },
+  platform: "node",
+};
 
-if (dev) {
-  await ctx.watch();
+const browserBuildOptions = {
+  ...commonBuildOptions,
+  entryPoints: {
+    "./client/dist/webview/DataViewer": "./client/src/webview/DataViewer.tsx",
+  },
+};
+
+if (process.env.npm_config_webviews || process.env.npm_config_client) {
+  const ctx = await esbuild.context(
+    process.env.npm_config_webviews ? browserBuildOptions : nodeBuildOptions,
+  );
+  await ctx.rebuild();
+
+  if (dev) {
+    await ctx.watch();
+  } else {
+    await ctx.dispose();
+  }
 } else {
-  await ctx.dispose();
+  const { result } = concurrently([
+    {
+      command: `npm run ${process.env.npm_lifecycle_event} --webviews`,
+      name: "browser",
+    },
+    {
+      command: `npm run ${process.env.npm_lifecycle_event} --client`,
+      name: "node",
+    },
+  ]);
+
+  await result.then(
+    () => {},
+    () => console.error("Assets failed to build successfully"),
+  );
 }